31/5

+++ lib/core/events/app_events.dart
@@ -1,12 +1,13 @@
 
 import 'package:supermarket/data/datasources/local/app_database.dart';
+import 'package:decimal/decimal.dart';
 
 abstract class AppEvent {}
 
 class SaleCreatedEvent extends AppEvent {
   final Sale sale;
   final List<SaleItem> items;
-  final double cogs;
+  final Decimal cogs;
   final String? userId;
 
   SaleCreatedEvent(this.sale, this.items, {required this.cogs, this.userId});
@@ -54,7 +55,7 @@ class PurchaseReturnCreatedEvent extends AppEvent {
 
 class CustomerPaymentEvent extends AppEvent {
   final String customerId;
-  final double amount;
+  final Decimal amount;
   final String paymentMethod;
   final String? note;
   final String paymentId;
@@ -72,7 +73,7 @@ class CustomerPaymentEvent extends AppEvent {
 
 class SupplierPaymentEvent extends AppEvent {
   final String supplierId;
-  final double amount;
+  final Decimal amount;
   final String paymentMethod;
   final String? note;
   final String paymentId;
@@ -89,7 +90,7 @@ class SupplierPaymentEvent extends AppEvent {
 }
 
 class CashTransactionEvent extends AppEvent {
-  final double amount;
+  final Decimal amount;
   final String type; // IN, OUT
   final String category;
   final String accountId;
